Use webpack in the deploy script

+++ buildscripts/deploy.js
@@ -1,17 +1,25 @@
 var exec = require('child_process').exec;
 var fs = require('fs');
-var build = require('./build.js');
+//var build = require('./build.js');
 var path = require('path');
 var rimraf = require('rimraf');
 var fs_extra = require('fs-extra');
+var webpack = require('webpack');
+var webpackConfig = require('../webpack.config');
 
 var tempDir = fs.mkdtempSync('../deploy');
 
-fs.createReadStream('index.html').pipe(fs.createWriteStream(path.join(tempDir, 'index.html')));
-build.copy(path.join(tempDir, "dist"));
+webpackConfig.output.path = path.join(tempDir, "dist");
 
-exec('git checkout gh-pages', function() {
-    //rimraf(destination, function() { });
-    fs_extra.copySync(tempDir, '.', {overwrite: true});
-    rimraf(tempDir, function() { });
+fs.createReadStream('index.html').pipe(fs.createWriteStream(path.join(tempDir, 'index.html')));
+
+webpack(webpackConfig, (err, stats) => {
+    if (err || stats.hasErrors()) {
+        console.log(err);
+        return;
+    }
+    exec('git checkout gh-pages', function () {
+        fs_extra.copySync(tempDir, '.', { overwrite: true });
+        rimraf(tempDir, function () { });
+    });
 });
